pyproject.toml
[tool.poetry]
name = "photoproject"
version = "0.1.0"
description = ""
authors = ["Yulia Kuznetsova <kooznitsa@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.1.6"
django-environ = "^0.12.0"
django-cors-headers = "^4.7.0"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
django-elasticsearch-dsl = "^8.0"
torchvision = "^0.21.0"
torch = "^2.6.0"
pillow = "^11.1.0"
googletrans = "^4.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
